+++ docs/history.rst
@@ -5,6 +5,7 @@ Release History
 ---------------
 - Renamed primary branch to "main"
 - Drop Python 2.7 support
+- Add type annotations
 
 `0.0.6`_ (8 Apr 2019)
 ---------------------