+++ docs/history.rst
@@ -6,6 +6,7 @@ Release History
 - Renamed primary branch to "main"
 - Drop Python 2.7 support
 - Add type annotations
+- Rewrite examples to use asyncio instead of the Tornado IOLoop
 
 `0.0.6`_ (8 Apr 2019)
 ---------------------