+++ docs/history.rst
@@ -1,12 +1,13 @@
 Release History
 ===============
 
-`Next Release`_
----------------
+`1.0.0`_ (2 Jul 2022)
+---------------------
 - Renamed primary branch to "main"
 - Drop Python 2.7 support
 - Add type annotations
 - Rewrite examples to use asyncio instead of the Tornado IOLoop
+- Drop support for older Tornado versions (6.0+ is required now)
 
 `0.0.6`_ (8 Apr 2019)
 ---------------------
@@ -29,7 +30,8 @@ Release History
 ----------------------
 - Initial alpha release containing a very simple implementation.
 
-.. _Next Release: https://github.com/dave-shawley/tornado-problem-details/compare/0.0.6...main
+.. _Next Release: https://github.com/dave-shawley/tornado-problem-details/compare/1.0.0...main
+.. _1.0.0: https://github.com/dave-shawley/tornado-problem-details/compare/0.0.6...1.0.0
 .. _0.0.6: https://github.com/dave-shawley/tornado-problem-details/compare/0.0.5...0.0.6
 .. _0.0.5: https://github.com/dave-shawley/tornado-problem-details/compare/0.0.4...0.0.5
 .. _0.0.4: https://github.com/dave-shawley/tornado-problem-details/compare/0.0.2...0.0.4